# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TraceHawk Phase 1 API contract",
    "version": "0.4.0",
    "description": "TraceHawk local replay, four explainable detectors and authenticated investigation API. Defaults include explicitly fictional indicators."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3100"
    }
  ],
  "paths": {
...
    "/api/v1/operations": {
      "get": {
        "operationId": "operations",
        "description": "Phase 3. Local-workspace operator access required; no quarantined payload bodies exposed.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "status",
                    "workers",
                    "partitions",
                    "queues"
                  ],
                  "properties": {
                    "status": {
                      "type": "string"
                    },
                    "workers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": [
                          "component",
                          "status",
                          "fresh",
                          "details"
                        ],
                        "properties": {
                          "component": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "fresh": {
                            "type": "boolean"
                          },
                          "details": {
                            "type": "object"
                          }
                        }
                      }
                    },
                    "partitions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": false,
                        "required": [
                          "partition",
                          "next_offset",
                          "ownership_epoch",
                          "owner_id",
                          "broker_high",
                          "lag",
                          "fresh"
                        ],
                        "properties": {
                          "partition": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "next_offset": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "ownership_epoch": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "owner_id": {
                            "anyOf": [
                              {
                                "type": "string"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          },
                          "broker_high": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "lag": {
                            "type": "integer",
                            "minimum": 0
                          },
                          "fresh": {
                            "anyOf": [
                              {
                                "type": "boolean"
                              },
                              {
                                "type": "null"
                              }
                            ]
                          }
                        }
                      }
                    },
                    "queues": {
                      "type": "object",
                      "additionalProperties": false,
                      "required": [
                        "pending_alert_updates",
                        "pending_deadletters",
                        "unreviewed_deadletters"
                      ],
                      "properties": {
                        "pending_alert_updates": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "pending_deadletters": {
                          "type": "integer",
                          "minimum": 0
                        },
                        "unreviewed_deadletters": {
                          "type": "integer",
                          "minimum": 0
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown scoped resource",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body too large",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tracehawk_session"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "message": {
            "type": "string",
            "maxLength": 512
          },
          "request_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          }
        }
      },
...
    }
  }
}